requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=9.5.0

# MiDaS Dependencies
timm>=0.9.0

# LLM API
openai>=1.0.0
google-generativeai>=0.5.0

# Data Processing
pandas>=2.0.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web Server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Web Interface (optional)
gradio>=3.40.0
streamlit>=1.25.0

# Development Tools
jupyter>=1.0.0
ipython>=8.0.0
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0

# Logging and Config
pyyaml>=6.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.2.0

# Image Processing
scikit-image>=0.20.0
albumentations>=1.3.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
requests>=2.31.0 